added new service to get data for public gallery

+++ backend/models/imageModel.js
@@ -2,6 +2,8 @@ import knex from "./knexfile.js";
 import path from 'path';
 import fs from 'fs';
 import sharp from 'sharp';
+import CommentModel from "./commentModel.js";
+import LikeModel from "./likeModel.js";
 
 export default class ImageModel {
     static async create(userId, overlayPath, backgoundPath) {
@@ -40,9 +42,23 @@ export default class ImageModel {
         return knex('Images').select('*').where('user_id', userId);
     }
 
-    static async deleteById(imageId) {
+    static async getAllByPage(page) {
+        let images = []
+        const limitPerPage = 5;
+        const offset = (page - 1) * limitPerPage;
+        images = await knex('Images').select('*').limit(limitPerPage).offset(offset);
+        for (let image of images) {
+            image.comments = await CommentModel.findAllByImageId(image.id);
+            console.log('image likes :', image.likes)
+            image.likes = await LikeModel.findAllByImageId(image.id);
+        }
+        return images
+    }
+
+    static async deleteById(imageId, userId) {
         if (!imageId) throw new Error('Missing required fields');
-        const image = await knex('Images').select('processed_path').where('id', imageId).first();
+        const image = await knex('Images').select('processed_path', 'user_id').where('id', imageId).first();
+        if (image.user_id !== userId) throw new Error(`Unauthorized`);
         if (image) {
             // Delete the file from storage
             fs.unlinkSync(image.processed_path);
